Fixed targeting systems for SVVP
Before, to prevent targeting identical targets on multiple machines with the
createprojecttarget request on a specific machine's target, we iterated on all
targets and check if it is the wanted target and target it, since systems are
more complicated type of target, this iteration doesn't work, and the right way
that should have been used from the start is to call CreateTarget from the
TargetFamily, now fixed, but we need the next commit which fixes the issue of
targeting identical targets on multiple machines and keep the SVVP support.

Also, when targeting a system in a project, other non system targets cant
be present in the same project and vice versa, added an appropiate
condition to prevent the vice versa part.

Signed-off-by: Bishara AbuHattoum <bishara@daynix.com>